fix: move course subject check to method and delete event listener

+++ backend/src/main/java/ar/net/ut/backend/subject/SubjectService.java
@@ -1,6 +1,8 @@
 package ar.net.ut.backend.subject;
 
+import ar.net.ut.backend.course.repository.CourseSubjectRepository;
 import ar.net.ut.backend.enums.ResourceType;
+import ar.net.ut.backend.exception.BackendException;
 import ar.net.ut.backend.exception.impl.ResourceAlreadyExistsException;
 import ar.net.ut.backend.exception.impl.ResourceNotFoundException;
 import ar.net.ut.backend.subject.dto.SubjectCreateDTO;
@@ -12,6 +14,7 @@ import ar.net.ut.backend.subject.event.SubjectUpdateEvent;
 import ar.net.ut.backend.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -21,11 +24,12 @@ import java.util.List;
 @RequiredArgsConstructor
 public class SubjectService {
 
+    private final UserService userService;
+
     private final SubjectRepository subjectRepository;
+    private final CourseSubjectRepository courseSubjectRepository;
 
     private final SubjectMapper subjectMapper;
-
-    private final UserService userService;
 
     private final ApplicationEventPublisher eventPublisher;
 
@@ -71,6 +75,13 @@ public class SubjectService {
     @Transactional
     public void deleteSubject(Long id) {
         Subject subject = getById(id);
+        if (courseSubjectRepository.existsBySubjectId(id)) {
+            throw new BackendException(
+                    "You must remove that subject from all courses in order to delete it",
+                    HttpStatus.CONFLICT,
+                    "SUBJECT_DELETE_COURSE_CONFLICT"
+            );
+        }
 
         subjectRepository.delete(subject);
         subjectRepository.unlinkSubjectFromCorrelatives(id);
